use HUDocumentViewAttributesHelper on in handlingunits package
https://github.com/metasfresh/metasfresh-webui-api/issues/269
