+++ src/main/java/de/metas/ui/web/view/descriptor/DocumentViewAttributesLayout.java
@@ -2,25 +2,10 @@ package de.metas.ui.web.view.descriptor;
 
 import java.util.List;
 
-import org.adempiere.model.InterfaceWrapperHelper;
-import org.adempiere.util.GuavaCollectors;
-import org.adempiere.util.Services;
-import org.compiere.model.I_M_Attribute;
-
 import com.google.common.base.MoreObjects;
+import com.google.common.collect.ImmutableList;
 
-import de.metas.device.adempiere.AttributesDevicesHub.AttributeDeviceAccessor;
-import de.metas.device.adempiere.IDevicesHubFactory;
-import de.metas.handlingunits.attribute.IAttributeValue;
-import de.metas.handlingunits.attribute.storage.IAttributeStorage;
-import de.metas.i18n.IModelTranslationMap;
-import de.metas.i18n.ITranslatableString;
-import de.metas.ui.web.devices.DeviceWebSocketProducerFactory;
-import de.metas.ui.web.devices.JSONDeviceDescriptor;
-import de.metas.ui.web.handlingunits.HUDocumentViewAttributesHelper;
-import de.metas.ui.web.window.descriptor.DocumentFieldWidgetType;
 import de.metas.ui.web.window.descriptor.DocumentLayoutElementDescriptor;
-import de.metas.ui.web.window.descriptor.DocumentLayoutElementFieldDescriptor;
 
 /*
  * #%L
@@ -46,22 +31,17 @@ import de.metas.ui.web.window.descriptor.DocumentLayoutElementFieldDescriptor;
 
 public final class DocumentViewAttributesLayout
 {
-	public static final DocumentViewAttributesLayout of(final IAttributeStorage attributeStorage)
+	public static final DocumentViewAttributesLayout of(final List<DocumentLayoutElementDescriptor> elements)
 	{
-		return new DocumentViewAttributesLayout(attributeStorage);
+		return new DocumentViewAttributesLayout(elements);
 	}
 
 	private final List<DocumentLayoutElementDescriptor> elements;
 
-	private DocumentViewAttributesLayout(final IAttributeStorage attributeStorage)
+	private DocumentViewAttributesLayout(final List<DocumentLayoutElementDescriptor> elements)
 	{
 		super();
-
-		final int warehouseId = attributeStorage.getM_Warehouse_ID();
-		elements = attributeStorage.getAttributeValues()
-				.stream()
-				.map(av -> createElement(av, warehouseId))
-				.collect(GuavaCollectors.toImmutableList());
+		this.elements = ImmutableList.copyOf(elements);
 	}
 
 	@Override
@@ -72,46 +52,6 @@ public final class DocumentViewAttributesLayout
 				.toString();
 	}
 
-	private static final DocumentLayoutElementDescriptor createElement(final IAttributeValue attributeValue, final int warehouseId)
-	{
-		final I_M_Attribute attribute = attributeValue.getM_Attribute();
-		final IModelTranslationMap attributeTrlMap = InterfaceWrapperHelper.getModelTranslationMap(attribute);
-		final ITranslatableString caption = attributeTrlMap.getColumnTrl(I_M_Attribute.COLUMNNAME_Name, attribute.getName());
-		final ITranslatableString description = attributeTrlMap.getColumnTrl(I_M_Attribute.COLUMNNAME_Description, attribute.getDescription());
-
-		final String attributeName = HUDocumentViewAttributesHelper.extractAttributeName(attributeValue);
-		final DocumentFieldWidgetType widgetType = HUDocumentViewAttributesHelper.extractWidgetType(attributeValue);
-
-		return DocumentLayoutElementDescriptor.builder()
-				.setCaption(caption)
-				.setDescription(description)
-				.setWidgetType(widgetType)
-				.addField(DocumentLayoutElementFieldDescriptor.builder(attributeName)
-						.setPublicField(true)
-						.addDevices(createDevices(attribute.getValue(), warehouseId)))
-				.build();
-	}
-
-	private static final List<JSONDeviceDescriptor> createDevices(final String attributeCode, final int warehouseId)
-	{
-		return Services.get(IDevicesHubFactory.class)
-				.getDefaultAttributesDevicesHub()
-				.getAttributeDeviceAccessors(attributeCode)
-				.stream(warehouseId)
-				.map(attributeDeviceAccessor -> createDevice(attributeDeviceAccessor))
-				.collect(GuavaCollectors.toImmutableList());
-
-	}
-
-	private static JSONDeviceDescriptor createDevice(final AttributeDeviceAccessor attributeDeviceAccessor)
-	{
-		final String deviceId = attributeDeviceAccessor.getPublicId();
-		return JSONDeviceDescriptor.builder()
-				.setDeviceId(deviceId)
-				.setCaption(attributeDeviceAccessor.getDisplayName())
-				.setWebsocketEndpoint(DeviceWebSocketProducerFactory.buildDeviceTopicName(deviceId))
-				.build();
-	}
 
 	public List<DocumentLayoutElementDescriptor> getElements()
 	{
